+++ changelog.md
@@ -5,6 +5,7 @@
 ### Improvements
 - LFG message now explains that a thread will be created automatically on game start
 - LFG buttons: Removed "Notify" added "Start" because notify is not a core feature but start is critical
+- Added instruction method for when boxscore token fails due to unrecognized thread
 
 ## [1.13.10] - 2026-9-20 Boxscore Bugs